requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
groq==0.4.1
langgraph==0.0.40
langchain-core==0.1.40
pydantic==2.5.0
